Add order accrual processing system with worker and database changes
- Add worker service to periodically check and update order accruals -
Change accrual fields from int to float64 in models and database - Add
new database migration for accrual column type change - Implement
accrual client for external system communication - Add new Makefile
target for running accrual service - Extend repository interface with
order processing methods

+++ internal/models/orderaccrual.go
@@ -17,5 +17,5 @@ const (
 type OrderAccrualResponse struct {
 	Order   string             `json:"order"`
 	Status  OrderAccrualStatus `json:"status"`
-	Accrual *int               `json:"accrual,omitempty"`
+	Accrual *float64           `json:"accrual,omitempty"` // Изменено с int на float64
 }
